switch to using es6 modules for better block bundling
Because I likely want to export multiple bundled blocks
for use in the code editor, ES6 modules are worth the
extra tooling.

now,
> caustics@0.0.1 build /Users/drew/code/codeblocks/caustics
> rollup -c -i blocks/projection.js -o blocks/dist/projection.js -f cjs will bundle just the relevant functions
from reflection.js into the code block source file.

Skip gl-matrix in the bundle because we have it on the NBWs
already anyway, and it's a commonJS module so I can't
tree-shake it.

Next up: rollup all the blocks in build step, not just projection.